EHCI: AMD extend handshake vaild time
On some AMD EHCI controllers, a Web camera issue
where due to their application software creating some traffic
activity to the controller, the controller may not respond to set
the Asynchronous Schedule Status bit immediately but may take up
to 40 ms to set it.
Linux USB driver/utility prematurely halts the USB controller if
it does not see the Asynchronous Schedule Status (bit 15) bit set
within a fixed amount of time after the Asynchronous Schedule Enable
(bit 5) is set.
So extend to about 40ms when the wait time of waiting for the response.

Signed-off-by: Huang Rui <ray.huang@amd.com>